[DPR-440] Component-ise data table
