+++ app/views/layout.html
@@ -59,7 +59,7 @@
 {% from "moj/components/primary-navigation/macro.njk"    import mojPrimaryNavigation %}
 {% from "moj/components/side-navigation/macro.njk"       import mojSideNavigation %}
 {% from "_components/card-group/macro.html"              import appCardGroup %}
-{% from "_components/data-table/macro.html"              import appDataTable %}
+{% from "../components/data-table/views/macro.html"      import appDataTable %}
 
 {% block head %}
   {% include "includes/head.html" %}
